Copy CMakeLists.txt from ./tlm on Windows
If the tlm project gets updated by the 'checkout_dependencies'
script and the tlm/CMakeLists.txt has been updated then we will
not automatically pull in the updated CMakeLists.txt as we do on
Linux since Windows does not use the top level 'Makefile'. This
patch means we will always copy in CMakeLists.txt in case it was
changed.

Change-Id: I3e9c749304f4999060cb18d3968e9ccc3b8abcdb